Gmcast Error: comparison doesn't meet irreflexive requirements
Refactor code to make problematic code path unit-testable.
- Extract static method `GMCast::compute_relay_set()` and make
  it public
- Move existing gmcast tests into separate check_nondet.cpp unit
- Create simple test for `GMCast::compute_relay_set()` to verify
  that it can be used

codership/galera-bugs#1044
