+++ gcomm/test/check_gcomm.cpp
@@ -33,6 +33,7 @@ static GCommSuite suites[] = {
     {"fair_send_queue", fair_send_queue_suite},
     {"util", util_suite},
     {"types", types_suite},
+    {"gmcast", gmcast_suite},
     {"evs2", evs2_suite},
     {"pc", pc_suite},
     {"", 0}
